Some changes and delete added
