+++ FoodOrderingApp-service/src/main/java/com/upgrad/FoodOrderingApp/service/businness/AddressBusinessService.java
@@ -18,6 +18,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.time.ZoneId;
 import java.time.ZonedDateTime;
+import java.util.List;
 
 @Service
 public class AddressBusinessService {
@@ -77,6 +78,34 @@ public class AddressBusinessService {
         return addressDao.save(customerAddressEntity);
     }
 
+    //To get Address By Customer
+    @Transactional(propagation = Propagation.REQUIRED)
+    public List<CustomerAddressEntity> getAddressByCustomer(final String authorizationToken) throws AuthorizationFailedException {
+
+        CustomerAuthTokenEntity customerAuth = customerDao.checkAuthToken(authorizationToken);
+
+        if (customerAuth.equals(null)) {
+            throw new AuthorizationFailedException("ATHR-001", "Customer is not Logged in.");
+        }
+        final ZonedDateTime customerSignOutTime = customerAuth.getLogoutAt();
+
+        if (customerSignOutTime != null && customerAuth != null) {
+            throw new AuthorizationFailedException("ATHR-002", "Customer is logged out. Log in again to access this endpoint.");
+        }
+
+        final ZonedDateTime customerSessionExpireTime = customerAuth.getExpiresAt();
+        ZonedDateTime currentTime = ZonedDateTime.now(ZoneId.systemDefault());
+        if (customerSessionExpireTime.compareTo(currentTime) < 0) {
+            throw new AuthorizationFailedException("ATHR-003", "Your session is expired. Log in again to access this endpoint.");
+        }
+
+        List<CustomerAddressEntity> customerAddressEntity = customerAddressDao.getAddressByCustomer(customerAuth.getCustomer());
+
+
+        return customerAddressEntity;
+
+    }
+
     public String deleteAddress(String addressId, String authorization) throws AuthorizationFailedException, AddressNotFoundException {
 
         //validate user
@@ -97,18 +126,18 @@ public class AddressBusinessService {
         if (customerSessionExpire.compareTo(currentTime) < 0) {
             throw new AuthorizationFailedException("ATHR-003", "Your session is expired. Log in again to access this endpoint.");
         }
-
+        if (addressEntity.getUuid().length() == 0){
+            throw new AddressNotFoundException("ANF-005", "Address id can not be empty");
+        }
+        if (addressEntity == null) {
+            throw new AddressNotFoundException("ANF-003", "No address by this id");
+        }
         if (customerAuthTokenEntity.getCustomer().getId().equals(customerAddressEntity.getCustomer().getId())) {
             return addressDao.deleteAddress(addressId);
         } else {
             throw new AuthorizationFailedException("ATHR-004", "You are not authorized to view/update/delete any one else's address");
         }
-        if (addressEntity.getUuid().length() == 0){
-            throw new AddressNotFoundException("ANF-005", "Address id can not be empty");
-            }
-        if (addressEntity == null) {
-            throw new AddressNotFoundException("ANF-003", "No address by this id");
-        }
+
     }
 
 }
